Scroll function moved to util, style updates
